+++ erp.backend/src/main/java/org/roko/erp/backend/services/PostedSalesOrderLineService.java
@@ -14,7 +14,7 @@ public interface PostedSalesOrderLineService {
 
     public List<PostedSalesOrderLine> list(PostedSalesOrder PostedSalesOrder, int page);
 
-    public int count(PostedSalesOrder postedSalesOrder);
+    public long count(PostedSalesOrder postedSalesOrder);
 
     public PostedSalesDocumentLineDTO toDTO(PostedSalesOrderLine postedSalesOrderLine);
 }
