+++ erp.backend/src/test/java/org/roko/erp/backend/controllers/ItemControllerTest.java
@@ -38,7 +38,7 @@ public class ItemControllerTest {
 
     private static final int TEST_PAGE = 12;
 
-    private static final int TEST_COUNT = 222;
+    private static final long TEST_COUNT = 222;
 
     @Captor
     private ArgumentCaptor<Item> itemArgumentCaptor;
