+++ erp.backend/src/test/java/org/roko/erp/backend/controllers/PaymentMethodControllerTest.java
@@ -32,7 +32,7 @@ public class PaymentMethodControllerTest {
 
     private static final int TEST_PAGE = 123;
 
-    private static final int TEST_COUNT = 222;
+    private static final long TEST_COUNT = 222;
 
     private static final String TEST_CODE = "test-code";
     private static final String TEST_NAME = "test-name";
